Image caption processing to utils
