+++ src/http/image-http.js
@@ -10,11 +10,7 @@ import * as userCore from '../core/user-core';
 import * as imageCore from '../core/image-core';
 import {assert} from '../validation';
 import {decodeBase64Image} from '../util/base64';
-import {
-  getRectangleTop,
-  getRectangleBottom,
-  getTextPositionFromCenter
-} from '../util/image-caption';
+import { processImageText } from '../util/image-caption';
 
 const logger = require('../util/logger')(__filename);
 const uuidV1 = require('uuid/v1');
@@ -80,67 +76,6 @@ function autoOrient(imageBuffer) {
         });
     } catch (err) {
       logger.error('Error in auto-orient:', err);
-      logger.error(err.stack);
-      reject(err);
-    }
-  });
-}
-
-function processImageText(imageBuffer, { imageText = '', imageTextPosition = 0.5 }) {
-  const BAR_HEIGHT = 60;
-  const BAR_COLOR = 'rgba(221, 73, 151, .6)';
-  const FONT_SIZE = 38;
-  const DEFAULT_WIDTH = 1024;
-  const DEFAULT_HEIGHT = 1024;
-  const TEXT_COLOR = '#FEFF77';
-  // const FONT_FAMILY = './CabinCondensed';
-
-  const textPosition = parseFloat(imageTextPosition);
-
-  return new Promise((resolve, reject) => {
-    try {
-      gm(imageBuffer)
-        .autoOrient()
-        .resize(DEFAULT_WIDTH, DEFAULT_HEIGHT)
-        .toBuffer('JPG', (resizeError, resizedBuffer) => {
-          if (resizeError) {
-            reject(resizeError);
-            return;
-          }
-
-          // Get size of resized image
-          gm(resizedBuffer)
-          .size((err, value) => {
-            const imageSize = !value || err
-              ? { width: DEFAULT_WIDTH, height: DEFAULT_HEIGHT }
-              : value;
-
-            // Start drawing on top of resized img
-            gm(resizedBuffer)
-            // Draw background for text
-            .fill(BAR_COLOR)
-            .drawRectangle(
-              0,
-              getRectangleTop(imageSize.height, textPosition, BAR_HEIGHT),
-              imageSize.width,
-              getRectangleBottom(imageSize.height, textPosition, BAR_HEIGHT)
-            )
-            // Draw text
-            .fill(TEXT_COLOR)
-            .fontSize(FONT_SIZE)
-            .drawText(
-              0,
-              getTextPositionFromCenter(imageSize.height, textPosition, BAR_HEIGHT),
-              imageText,
-              'Center'
-            )
-            .toBuffer('JPG', (error, resultBuffer) => {
-              error ? reject(error) : resolve(resultBuffer);
-            });
-          })
-      })
-    } catch (err) {
-      logger.error('Error in drawing image caption:', err);
       logger.error(err.stack);
       reject(err);
     }
